telemetria/requirements.txt
pika        # Para RabbitMQ
schedule    # Para tareas periódicas (actualizar posiciones)
python-dotenv # acceso a .env
python-consul